pyproject.toml
[project]
name = "telegram-calendar-bot"
version = "0.1.0"
description = "Telegram bot that creates Google Calendar events from natural language text"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiogram>=3.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-api-python-client>=2.0",
    "dateparser>=1.2.0",
    "cryptography>=42.0",
    "structlog>=24.0",
    "sentry-sdk[fastapi]>=2.0",
    "apscheduler>=3.10,<4.0",
    "tenacity>=8.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "coverage>=7.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0",
    "types-dateparser>=1.2.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["B008"]  # FastAPI Depends() in function defaults is an intentional pattern

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "google.*",
    "google_auth_oauthlib.*",
    "googleapiclient.*",
    "apscheduler.*",
    "dateparser.*",
    "aiogram.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
